% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_my_predictions.R
\name{add_my_predictions}
\alias{add_my_predictions}
\title{add_my_predictions}
\usage{
add_my_predictions(
  data = testing_data,
  model_1 = model_1,
  model_2 = model_1,
  model_3 = model_1
)
}
\arguments{
\item{model_3}{}
}
\value{
Adds predictions to testing_data set
}
\description{
add_my_predictions
}
