% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_songs.R
\name{get_my_songs}
\alias{get_my_songs}
\title{get_my_songs}
\usage{
get_my_songs(student_id)
}
\arguments{
\item{student_id}{must be 8 to 10 digits long, numeric}
}
\value{
Dataset of K-pop and Jazz songs
}
\description{
get_my_songs
}
\examples{
get_my_songs(1002345678)
}
