% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_starship.R
\name{get_my_starship}
\alias{get_my_starship}
\title{get_my_starship}
\usage{
get_my_starship(student_id)
}
\arguments{
\item{student_id}{must be 8 to 10 digits long, numeric}
}
\value{
Dataset about starship
}
\description{
get_my_starship
}
\examples{
get_my_starship(1002345678)
}
