% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shipname.R
\name{get_shipname}
\alias{get_shipname}
\title{get_shipname}
\usage{
get_shipname(student_id)
}
\arguments{
\item{student_id}{must be 8 to 10 digits long, numeric}
}
\value{
Single character element with the name of the space ship, created as 'SS' for space ship plus a made up word.
}
\description{
get_shipname
}
\examples{
get_shipname(1002345678)

}
